policies: Restrict negative access removal
Why:

* A user could escalate privileges by removing the negative accesses
applying to itself

+++ wazo_auth/plugins/http/policies/http.py
@@ -58,6 +58,13 @@ class _Policy(_BasePolicyRessource):
         return policy_full_schema.dump(policy), 200
 
     def _delete(self, policy_uuid, tenant_uuids):
+        token = Token.from_headers()
+        policy = self.policy_service.get(policy_uuid, tenant_uuids)
+        access_check = AccessCheck(token.auth_id, token.session_uuid, token.acl)
+        for access in policy.acl:
+            if not access_check.may_remove_access(access):
+                raise Unauthorized(token.token, required_access=access)
+
         self.policy_service.delete(policy_uuid, tenant_uuids)
         return '', 204
 
@@ -72,6 +79,11 @@ class _Policy(_BasePolicyRessource):
         access_check = AccessCheck(token.auth_id, token.session_uuid, token.acl)
         for access in body['acl']:
             if not access_check.may_add_access(access):
+                raise Unauthorized(token.token, required_access=access)
+
+        policy = self.policy_service.get(policy_uuid, tenant_uuids)
+        for access in set(policy.acl) - set(body['acl']):
+            if not access_check.may_remove_access(access):
                 raise Unauthorized(token.token, required_access=access)
 
         body['tenant_uuids'] = tenant_uuids
@@ -118,6 +130,10 @@ class PolicySlug(_Policy):
 
 class _PolicyAccess(_BasePolicyRessource):
     def _delete(self, policy_uuid, access, tenant_uuids):
+        token = Token.from_headers()
+        access_check = AccessCheck(token.auth_id, token.session_uuid, token.acl)
+        if not access_check.may_remove_access(access):
+            raise Unauthorized(token.token, required_access=access)
         self.policy_service.delete_access(policy_uuid, access, tenant_uuids)
         return '', 204
